refactor(web/runner): Improve OpenDC web runner implementation
This change contains a rewrite of the OpenDC web runner implementation,
which now supports terminating simulations when exceeding a deadline, as
well as executing multiple simulation jobs at the same time.

Furthermore, we have extracted the runner from the command line
interface, so that we can offer this functionality as a library in the
future.
